Refactor remaining components
