+++ app/components/Template.js
@@ -1,15 +1,22 @@
+'use client';
 import Markdown from 'react-markdown';
 import { Paper } from '@mui/material';
+import { useContext } from 'react';
+import { AppContext } from '../contexts/appDetails';
 import './Template.css';
 
-export default function Template({ file }) {
+
+export default function Template() {
+
+  const { file } = useContext(AppContext);
+
   return (
     <>
-      {file && (
+      {file.template && (
         <>
           <p>Template</p>
           <Paper sx={{ px: 4, py: 2 }}>
-            <Markdown>{file}</Markdown>
+            <Markdown>{file.template}</Markdown>
           </Paper>
         </>
       )}
